Resolve some script loading issues
