+++ checkov/terraform/module_loading/__init__.py
@@ -1,0 +1,1 @@
+from checkov.terraform.module_loading.loaders import *
