+++ checkov/terraform/module_loading/loaders/local_path_loader.py
@@ -1,7 +1,8 @@
 import os
 from typing import Optional
 
-from checkov.terraform.module_loading.loader import ModuleLoader, ModuleContent
+from checkov.terraform.module_loading.loader import ModuleLoader
+from checkov.terraform.module_loading.content import ModuleContent
 
 
 class LocalPathLoader(ModuleLoader):
